requirements.txt
langchain
faiss-cpu
openai
sentence-transformers
pypdf
gradio